Create tests

+++ src/main/java/com/example/weather_forecast_service/dto/WeatherInformation.java
@@ -13,5 +13,5 @@ public class WeatherInformation implements Serializable {
 
     private Integer temperature;
 
-    private Integer humidity;
+    private Long humidity;
 }
